[MOV] partner_autocomplete: move autocomplete feature directly into iap_mail
Autocomplete on partner is moved direclty into iap_mail module. This way it
can be called independently from business flows. Autocomplete module still
holds interface for asynchronous update of partners as well as automatic
enrich of companies.
